#0 | @rashjredmund | fixed Dockerfile failing to build TypeScript problems, and updated folder structures. Also updated Dockerfiles and abit of the docker-compos.yaml file
