+++ README.md
@@ -23,3 +23,7 @@ RB Clean is a simple laundry pickup service service
     "message": "" // optional
   }
 ```
+
+### The Amazing Docker Guide (by [Rash](https://github.com/rashjredmund) for the People!)
+
+- k
