+++ docker-compose.yaml
@@ -31,8 +31,6 @@ services:
     ports:
       - "9000:3000"
     restart: always
-    environment:
-      PORT: 3000
 
   rb-clean-profile:
     build: ./profile/
@@ -41,5 +39,4 @@ services:
       - "9001:3001"
     restart: always
     environment:
-      PORT: 3001
       DATABASE_URL: "" # add profile service db url here
