+++ profile/Dockerfile
@@ -1,9 +1,9 @@
 # build stage
-FROM node:16-alpine AS build
+FROM node:18-alpine AS build_stage
 
 WORKDIR /app
 
-COPY package*.json /app
+COPY package*.json .
 
 RUN npm install
 
@@ -12,15 +12,17 @@ COPY . .
 RUN npm run build
 
 # production stage
-FROM node:16-alpine AS production
+FROM node:18-alpine AS production
 
 WORKDIR /app
 
 COPY package*.json .
 
-# npm ci to "clean install" from package-lock.json
-RUN npm ci --only=production
+COPY --from=build_stage /app/build/ ./build
 
-COPY --from=build /app/build/ ./build
+# npm ci to "clean install" from package-lock.json
+RUN npm i --omit=dev
+
+ENV PORT="3001"
 
 CMD ["node", "build/app.js"]
